Put resource labels on replica-binary for root vol
When we create a volume with a backing file, we add a replica-binary
container from which the replica does a volumes_from. Rancher schedules
containers that are the target of a volumes_from first when scheduling a
deployment unit. Normally, scheduling labels are copied amongst all
containers in a deployment unit, so deployment order doesn't matter.

But these resource labels are purposefully not scheduling labels so that
they do not get copied to all containers and thus throw the resource
based scheduler off.

So with these labels, order matters and they must be on the first
container in the deployment unit to get deployed. So, this change puts
the labels on the replica-binary container instead of the replica
container when it is present.