Fix restore error
Pass the process id as the identifier for the restore process so that
it can be uniquely identified if the restore is retried. Also use an
explicit volumeName parameter for constructing the volume api URL. This
is necessary because the volume embedded inside backup is where the
backup originated but the target of the restore may be a different
volume.